backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
google-auth
gspread
obspy==1.2.2
pandas==1.3.3
python-dotenv
pytz==2021.3
SQLAlchemy==1.3.20
uWSGI==2.0.19.1